requirements.txt
aniso8601==8.0.0
appdirs==1.4.3
attrs==19.3.0
CacheControl==0.12.6
certifi==2019.9.11
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
diskcache==5.0.3
distlib==0.3.0
distro==1.4.0
editdistance==0.5.3
Flask==1.1.2
Flask-Cors==3.0.8
flask-restplus==0.13.0
gevent==20.6.2
greenlet==0.4.16
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
importlib-metadata==1.7.0
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
jsonschema==3.2.0
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
numpy==1.19.0
packaging==20.3
pep517==0.8.2
progress==1.5
pyenchant==3.1.1
pyparsing==2.4.6
PyPDF2==1.26.0
pyrsistent==0.16.0
pytoml==0.1.21
pytz==2020.1
regex==2020.6.8
requests==2.22.0
retrying==1.3.3
six==1.12.0
soupsieve==1.9.3
tika==1.19
toml==0.10.1
tqdm==4.47.0
urllib3==1.25.5
webencodings==0.5.1
Werkzeug==0.16.1
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
